split the contexts into Wallet and User
