+++ src/config.ts
@@ -2,3 +2,4 @@
 export const DEFAULT_WAB_URL = 'https://wab.babbage.systems';
 export const DEFAULT_STORAGE_URL = 'https://storage.babbage.systems';
 export const DEFAULT_CHAIN = 'main';
+export const ADMIN_ORIGINATOR = 'admin.com';
